requirements.txt
streamlit==1.25.0
python-docx==0.8.11
docx2txt==0.8
pandas==2.2.0